reorganize commcare_cloud.py into a number of files
