+++ commcare-cloud/commcare_cloud/commands/ansible/ansible_playbook.py
@@ -1,0 +1,204 @@
+# coding=utf-8
+import os
+import subprocess
+from six.moves import shlex_quote
+from clint.textui import puts, colored
+from commcare_cloud.cli_utils import ask, has_arg, check_branch
+from commcare_cloud.commands.ansible.helpers import AnsibleContext, DEPRECATED_ANSIBLE_ARGS, \
+    get_common_ssh_args
+from commcare_cloud.commands.shared_args import arg_inventory_group, arg_skip_check, arg_quiet, \
+    arg_branch, arg_stdout_callback
+from commcare_cloud.parse_help import add_to_help_text, filtered_help_message
+from commcare_cloud.environment import ANSIBLE_DIR, get_inventory_filepath, get_vault_vars_filepath, \
+    get_public_vars_filepath, get_public_vars
+
+
+class AnsiblePlaybook(object):
+    command = 'ansible-playbook'
+    help = (
+        "Run a playbook as you would with ansible-playbook, "
+        "but with boilerplate settings already set based on your <environment>. "
+        "By default, you will see --check output and then asked whether to apply. "
+    )
+
+    @staticmethod
+    def make_parser(parser):
+        arg_skip_check(parser)
+        arg_quiet(parser)
+        arg_branch(parser)
+        arg_stdout_callback(parser)
+        parser.add_argument('playbook', help=(
+            "The ansible playbook .yml file to run."
+        ))
+        add_to_help_text(parser, "\n{}\n{}".format(
+            "The ansible-playbook options below are available as well",
+            filtered_help_message(
+                "ansible-playbook -h",
+                below_line='Options:',
+                above_line=None,
+                exclude_args=DEPRECATED_ANSIBLE_ARGS + [
+                    '--help',
+                    '--diff',
+                    '--check',
+                    '-i',
+                    '--ask-vault-pass',
+                    '--vault-password-file',
+                ],
+            )
+        ))
+
+    @staticmethod
+    def run(args, unknown_args):
+        ansible_context = AnsibleContext(args)
+        check_branch(args)
+        public_vars = get_public_vars(args.environment)
+        ask_vault_pass = public_vars.get('commcare_cloud_use_vault', True)
+
+        def ansible_playbook(environment, playbook, *cmd_args):
+            cmd_parts = (
+                'ansible-playbook',
+                os.path.join(ANSIBLE_DIR, '{playbook}'.format(playbook=playbook)),
+                '-i', get_inventory_filepath(environment),
+                '-e', '@{}'.format(get_vault_vars_filepath(environment)),
+                '-e', '@{}'.format(get_public_vars_filepath(environment)),
+                '--diff',
+            ) + cmd_args
+
+            if not has_arg(unknown_args, '-u', '--user'):
+                cmd_parts += ('-u', 'ansible')
+
+            if not has_arg(unknown_args, '-f', '--forks'):
+                cmd_parts += ('--forks', '15')
+
+            if has_arg(unknown_args, '-D', '--diff') or has_arg(unknown_args, '-C', '--check'):
+                puts(colored.red("Options --diff and --check not allowed. Please remove -D, --diff, -C, --check."))
+                puts("These ansible-playbook options are managed automatically by commcare-cloud and cannot be set manually.")
+                exit(2)
+
+            if ask_vault_pass:
+                cmd_parts += ('--vault-password-file=/bin/cat',)
+
+            cmd_parts += get_common_ssh_args(public_vars)
+            cmd = ' '.join(shlex_quote(arg) for arg in cmd_parts)
+            print(cmd)
+            p = subprocess.Popen(cmd, stdin=subprocess.PIPE, shell=True, env=ansible_context.env_vars)
+            if ask_vault_pass:
+                p.communicate(input='{}\n'.format(ansible_context.get_ansible_vault_password()))
+            else:
+                p.communicate()
+            return p.returncode
+
+        def run_check():
+            return ansible_playbook(args.environment, args.playbook, '--check', *unknown_args)
+
+        def run_apply():
+            return ansible_playbook(args.environment, args.playbook, *unknown_args)
+
+        exit_code = 0
+
+        if args.skip_check:
+            user_wants_to_apply = ask('Do you want to apply without running the check first?',
+                                      quiet=args.quiet)
+        else:
+            exit_code = run_check()
+            if exit_code == 1:
+                # this means there was an error before ansible was able to start running
+                exit(exit_code)
+                return  # for IDE
+            elif exit_code == 0:
+                puts(colored.green(u"✓ Check completed with status code {}".format(exit_code)))
+                user_wants_to_apply = ask('Do you want to apply these changes?',
+                                          quiet=args.quiet)
+            else:
+                puts(colored.red(u"✗ Check failed with status code {}".format(exit_code)))
+                user_wants_to_apply = ask('Do you want to try to apply these changes anyway?',
+                                          quiet=args.quiet)
+
+        if user_wants_to_apply:
+            exit_code = run_apply()
+            if exit_code == 0:
+                puts(colored.green(u"✓ Apply completed with status code {}".format(exit_code)))
+            else:
+                puts(colored.red(u"✗ Apply failed with status code {}".format(exit_code)))
+
+        exit(exit_code)
+
+
+class _AnsiblePlaybookAlias(object):
+    @staticmethod
+    def make_parser(parser):
+        arg_skip_check(parser)
+        arg_quiet(parser)
+        arg_branch(parser)
+        arg_stdout_callback(parser)
+
+
+class UpdateConfig(_AnsiblePlaybookAlias):
+    command = 'update-config'
+    help = (
+        "Run the ansible playbook for updating app config "
+        "such as django localsettings.py and formplayer application.properties."
+    )
+
+    @staticmethod
+    def run(args, unknown_args):
+        args.playbook = 'deploy_localsettings.yml'
+        unknown_args += ('--tags=localsettings',)
+        AnsiblePlaybook.run(args, unknown_args)
+
+
+class AfterReboot(_AnsiblePlaybookAlias):
+    command = 'after-reboot'
+    help = (
+        "Bring a just-rebooted machine back into operation. "
+        "Includes mounting the encrypted drive."
+    )
+
+    @staticmethod
+    def make_parser(parser):
+        _AnsiblePlaybookAlias.make_parser(parser)
+        arg_inventory_group(parser)
+
+    @staticmethod
+    def run(args, unknown_args):
+        args.playbook = 'deploy_stack.yml'
+        args.skip_check = True
+        unknown_args += ('--tags=after-reboot', '--limit', args.inventory_group)
+        AnsiblePlaybook.run(args, unknown_args)
+
+
+class RestartElasticsearch(_AnsiblePlaybookAlias):
+    command = 'restart-elasticsearch'
+    help = (
+        "Do a rolling restart of elasticsearch."
+    )
+
+    @staticmethod
+    def run(args, unknown_args):
+        args.playbook = 'es_rolling_restart.yml'
+        if not ask('Have you stopped all the elastic pillows?', strict=True, quiet=args.quiet):
+            exit(0)
+        puts(colored.yellow(
+            "This will cause downtime on the order of seconds to minutes,\n"
+            "except in a few cases where an index is replicated across multiple nodes."))
+        if not ask('Do a rolling restart of the ES cluster?', strict=True, quiet=args.quiet):
+            exit(0)
+        AnsiblePlaybook.run(args, unknown_args)
+
+
+class BootstrapUsers(_AnsiblePlaybookAlias):
+    command = 'bootstrap-users'
+    help = (
+        "Add users to a set of new machines as root. "
+        "This must be done before any other user can log in."
+    )
+
+    @staticmethod
+    def run(args, unknown_args):
+        args.playbook = 'deploy_stack.yml'
+        public_vars = get_public_vars(args.environment)
+        root_user = public_vars.get('commcare_cloud_root_user', 'root')
+        unknown_args += ('--tags=users', '-u', root_user)
+        if not public_vars.get('commcare_cloud_pem'):
+            unknown_args += ('--ask-pass',)
+        AnsiblePlaybook.run(args, unknown_args)
